+++ pyobjc-core/PyObjCTest/test_lazy_import.py
@@ -554,7 +554,7 @@ class TestLazyImport(TestCase):
         self.assertIn("magic instance", repr(mod.kCFAllocatorDefault))
 
         # XXX: These need to be in a different test file
-        self.assertTrue(getattr(mod.kCFAllocatorDefault, "__is_magic")())  # noqa: B009
+        self.assertTrue(mod.kCFAllocatorDefault.__pyobjc_magic_coookie__)
         self.assertTrue(mod.kCFAllocatorDefault == mod.kCFAllocatorDefault)
         self.assertFalse(mod.kCFAllocatorDefault != mod.kCFAllocatorDefault)
         self.assertTrue(mod.kCFAllocatorDefault != mod.CFBagRef)